fix again defining user pydantic model
